Crate lattices

Module containing lattice traits and implementations.

Convention: Generic parameters that are full words (e.g. Other, Key, Val) are lattice types. Conversely, Generic parameters that are single letters or acronyms (e.g. K, T) are scalar non-Lattice types.

Modules

bottom

Gives a default representation that compares as less than to a lattice.

cc_traits

Re-export of the cc_traits crate with SimpleKeyedRef added.

collections

Simple singleton or array collection with [cc_traits] implementations.

dom_pair

Dominating pair compound lattice.

fake

A fake lattice that will runtime panic if a merge is attempted.

map_union

Map-union compound lattice.

ord

Totally-ordered lattices, Max and Min.

pair

Pair compound lattice.

set_union

Set-union compound lattice.

Traits

Compare

Compare the partial order of two lattices.

ConvertFrom

Same as From but for lattices.

Merge

Trait for lattice merge (least upper bound).

NaiveCompare

Naive lattice compare, based on the Merge::merge function.