Module lattices::cc_traits

Re-export of the cc_traits crate with SimpleKeyedRef added.

Macros

covariant_item_mut

Automatically defines the CollectionMut::upcast_item_mut function using the covariance of the ItemMut<'a> type with regards to 'a.

covariant_item_ref

Automatically defines the CollectionRef::upcast_item_ref function using the covariance of the ItemRef<'a> type with regards to 'a.

covariant_key_ref

Automatically defines the KeyedRef::upcast_item_ref function using the covariance of the KeyRef<'a> type with regards to 'a.

Traits

Back

Collection exposing a reference to its back element.

BackMut

Collection exposing a mutable reference to its back element.

Capacity

Collection with known capacity.

CapacityMut

Collection with mutable capacity.

Clear

Clearable collection.

Collection

Abstract collection.

CollectionMut

Abstract collection that can be mutably referenced.

CollectionRef

Abstract collection that can be immutably referenced.

Deque

Immutable double-ended queue.

DequeMut

Mutable double-ended queue.

Front

Collection exposing a reference to its front element.

FrontMut

Collection exposing a mutable reference to its front element.

Get

Queryable collection.

GetKeyValue

Queryable map.

GetKeyValueMut

Mutably queryable map.

GetMut

Mutably queryable collection.

Insert

Mutable collection where new elements can be inserted.

Iter

Iterable collection.

IterMut

Mutably iterable collection.

Keyed

Abstract keyed collection.

KeyedRef

Abstract keyed collection whose key can be referenced.

Len

Sized collection.

Map

Imutable map data structure.

MapInsert

Mutable map where new new key-value pairs can be inserted.

MapIter

MapIterMut

MapMut

Mutable map data structure.

PopBack

Mutable collection where elements can be popped from the back.

PopFront

Mutable collection where elements can be popped from the front.

PushBack

Mutable collection where new elements can be pushed on the back.

PushFront

Mutable collection where new elements can be pushed on the front.

Remove

Mutable collection where elements can be removed from.

Reserve

Collection that can extend their capacity.

Set

Imutable set data structure.

SetMut

Mutable set data structure.

SimpleCollectionMut

Collection where each item reference can be converted into a standard “simple” rust reference.

SimpleCollectionRef

Collection where each item reference can be converted into a standard “simple” rust reference.

SimpleKeyedRef

https://github.com/timothee-haudebourg/cc-traits/pull/8

Slab

Imutable slab data structure.

SlabMut

Mutable slab data structure.

Stack

Immutable stack data structure.

StackMut

Mutable stack data structure.

Vec

Immutable array data structure (conventionally nammed “Vec”).

VecDeque

Immutable indexable deque.

VecDequeMut

Mutable indexable deque.

VecMut

Mutable Vec data structure.

WithCapacity

Collection that can be created with a minimum given capacity.