Trait lattices::cc_traits::CollectionRef

pub trait CollectionRef: Collection {
    type ItemRef<'a>: Clone + Deref<Target = Self::Item>
       where Self: 'a;

    // Required method
    fn upcast_item_ref<'short, 'long>(
        r: Self::ItemRef<'long>
    ) -> Self::ItemRef<'short>
       where 'long: 'short,
             Self: 'long;
}

Abstract collection that can be immutably referenced.

Required Associated Types

type ItemRef<'a>: Clone + Deref<Target = Self::Item> where Self: 'a

Type of references to items of the collection.

Required Methods

fn upcast_item_ref<'short, 'long>( r: Self::ItemRef<'long> ) -> Self::ItemRef<'short>where 'long: 'short, Self: 'long,

Changes an item reference into a shorter lived reference.

Item references are covariant with regard to the defined lifetime parameter 'a. Since this cannot be directly expressed by the type system, this associated function allows one to explicitly shorten the reference’s lifetime.

You can use the covariant_item_ref! macro to automatically implement this function.