Trait lattices::cc_traits::PushBack

pub trait PushBack: Collection {
    type Output;

    // Required method
    fn push_back(&mut self, element: Self::Item) -> Self::Output;
}

Mutable collection where new elements can be pushed on the back.

Required Associated Types

type Output

The output of the push function.

Required Methods

fn push_back(&mut self, element: Self::Item) -> Self::Output

Push a new element on the back of the collection.