Struct lattices::collections::VecMap

pub struct VecMap<K, V> {
    pub keys: Vec<K>,
    pub vals: Vec<V>,
}

A Vec-wrapper representing a naively implemented map.

Fields
keys: Vec<K>

Keys, should be the same length as and correspond 1:1 to vals.

vals: Vec<V>

Vals, should be the same length as and correspond 1:1 to keys.

Implementations

impl<K, V> VecMap<K, V>

pub fn new(keys: Vec<K>, vals: Vec<V>) -> Self

Create a new VecMap from the separate keys and vals vecs.

Panics if keys and vals are not the same length.