Module lattices::ght

GHT from the Wang/Willsey/Suciu Freejoin work

Modules

colt

COLT from Wang/Willsey/Suciu

lattice

Lattice traits for GHT

macros

Macros for GHT

test

Tests for the GHT code

Structs

GhtInner

internal node of a HashTrie

GhtLeaf

leaf node of a HashTrie

Traits

GeneralizedHashTrieNode

The GeneralizedHashTrieNode trait captures the properties of nodes in a Ght.

GhtGet

A trait for the get and iter methods from Wang/Willsey/Suciu, which work differently on leaves than internal nodes

GhtPrefixIter

A trait to iterate through the items in a Ght based on a prefix of the schema.