Trait lattices::ght::GhtGet

pub trait GhtGet: GeneralizedHashTrieNode {
    type Get: GeneralizedHashTrieNode<Schema = Self::Schema>;

    // Required methods
    fn get<'a>(&'a self, head: &Self::Head) -> Option<&'a Self::Get>;
    fn get_mut<'a>(&'a mut self, head: &Self::Head) -> Option<&'a mut Self::Get>;
    fn iter(&self) -> impl Iterator<Item = Self::Head>;
    fn iter_tuples(
        &self,
    ) -> impl Iterator<Item = <Self::Schema as VariadicExt>::AsRefVar<'_>>;
}

A trait for the get and iter methods from Wang/Willsey/Suciu, which work differently on leaves than internal nodes

Required Associated Types

type Get: GeneralizedHashTrieNode<Schema = Self::Schema>

Type returned by Self::get.

Required Methods

fn get<'a>(&'a self, head: &Self::Head) -> Option<&'a Self::Get>

On an Inner node, retrieves the value (child) associated with the given “head” key. returns an Option containing a reference to the value if found, or None if not found. On a Leaf node, returns None.

fn get_mut<'a>(&'a mut self, head: &Self::Head) -> Option<&'a mut Self::Get>

get, but mutable output

fn iter(&self) -> impl Iterator<Item = Self::Head>

Iterator for the “head” keys (from inner nodes) or nothing (from leaf nodes).

fn iter_tuples( &self, ) -> impl Iterator<Item = <Self::Schema as VariadicExt>::AsRefVar<'_>>

Iterator for the tuples (from leaf nodes) or nothing (from inner nodes).

Object Safety

This trait is not object safe.