Trait lattices::ght_lazy::GhtTakeLeaf

pub trait GhtTakeLeaf: GeneralizedHashTrieNode + Sealed {
    // Required methods
    fn take_containing_leaf(
        &mut self,
        row: <Self::Schema as VariadicExt>::AsRefVar<'_>,
    ) -> Option<GhtLeaf<Self::Schema, Self::ValType, Self::Storage>>;
    fn merge_leaf(
        &mut self,
        row: <Self::Schema as VariadicExt>::AsRefVar<'_>,
        leaf: GhtLeaf<Self::Schema, Self::ValType, Self::Storage>,
    ) -> bool;
}

Trait for taking a matching leaf out of a trie, as part of force method for COLTs

Required Methods

fn take_containing_leaf( &mut self, row: <Self::Schema as VariadicExt>::AsRefVar<'_>, ) -> Option<GhtLeaf<Self::Schema, Self::ValType, Self::Storage>>

take a matching leaf out of a trie, as part of the force method for COLTs

fn merge_leaf( &mut self, row: <Self::Schema as VariadicExt>::AsRefVar<'_>, leaf: GhtLeaf<Self::Schema, Self::ValType, Self::Storage>, ) -> bool

Merges a leaf into the hash trie

Object Safety

This trait is not object safe.