lattices::map_union_with_tombstones

Struct MapUnionWithTombstones

pub struct MapUnionWithTombstones<Map, TombstoneSet> { /* private fields */ }

Map-union-with-tombstones compound lattice.

When a key is deleted from the map-union-with-tombstones lattice, it is removed from the underlying map and placed into the tombstones set.

This forms the first invariant for this data structure. A key should appear either nowhere, in map or in tombstones. but never in map and tombstones at the same time.

merging is done by merging the underlying map and then merging the tombstones set, then doing map = map - tombstones.

The implementation of tombstones can be any set-like thing. This allows a user to plug in their own set-like implementation. For example, if the user knows that keys will be created and deleted strictly sequentially, then they could create a highly optimized set implementation which would just be a single integer, correpsonding to the current key value that the set is up to. Queries for keys below that integer would return true, queries for keys above it would return false.

Implementations

impl<Map, TombstoneSet> MapUnionWithTombstones<Map, TombstoneSet>

pub fn new(map: Map, tombstones: TombstoneSet) -> Self

Create a new MapUnionWithTombstones from a Map and a TombstoneSet.

pub fn new_from( map: impl Into<Map>, tombstones: impl Into<TombstoneSet>, ) -> Self

Create a new MapUnionWithTombstones from an Into<Map> and an Into<TombstoneSet>.

pub fn as_reveal_ref(&self) -> (&Map, &TombstoneSet)

Reveal the inner value as a shared reference.

pub fn as_reveal_mut(&mut self) -> (&mut Map, &mut TombstoneSet)

Reveal the inner value as an exclusive reference.

pub fn into_reveal(self) -> (Map, TombstoneSet)

Gets the inner by value, consuming self.