Struct lattices::ord::Min

#[repr(transparent)]
pub struct Min<T>(pub T);

A totally ordered min lattice. Merging takes the smaller value.

Tuple Fields
0: T
Implementations

impl<T> Min<T>

pub fn new(val: T) -> Self

Create a new Min lattice instance from a T.

pub fn new_from(val: impl Into<T>) -> Self

Create a new Min lattice instance from an Into<T> value.