Struct lattices::set_union_with_tombstones::SetUnionWithTombstones

pub struct SetUnionWithTombstones<Set, TombstoneSet> { /* private fields */ }

Set-union lattice with tombstones.

When an item is deleted from the SetUnionWithTombstones, it is removed from set and added to tombstones. This also is an invariant, if an item appears in tombstones it must not also be in set.

Merging set-union lattices is done by unioning the keys of both the (set and tombstone) sets, and then performing set = set - tombstones, to preserve the above invariant.

TODO: I think there is even one more layer of abstraction that can be made here. this ‘SetUnionWithTombstones’ can be turned into a kind of interface, because there are multiple underlying implementations. This implementation is two separate sets. Another implementation could be MapUnion<Key, WithTop<()>>, this would be less hash lookups, but maybe gives you less options for cool storage tricks. This implementation with two separate sets means that the actual set implementation can be decided for both the regular set and the tombstone set. Lots of opportunities there for cool storage tricks.

Implementations

impl<Set, TombstoneSet> SetUnionWithTombstones<Set, TombstoneSet>

pub fn new(set: Set, tombstones: TombstoneSet) -> Self

Create a new SetUnionWithTombstones from a Set and TombstoneSet.

pub fn new_from( set: impl Into<Set>, tombstones: impl Into<TombstoneSet> ) -> Self

Create a new SetUnionWithTombstones from an Into<Set> and an Into<TombstonesSet>.

pub fn as_reveal_ref(&self) -> (&Set, &TombstoneSet)

Reveal the inner value as a shared reference.

pub fn as_reveal_mut(&mut self) -> (&mut Set, &mut TombstoneSet)

Reveal the inner value as an exclusive reference.

pub fn into_reveal(self) -> (Set, TombstoneSet)

Gets the inner by value, consuming self.