Module lattices::test

Helper test utils to test lattice implementation correctness.

Functions

cartesian_power

Returns an iterator of N-length arrays containing all possible permutations (with replacement) of items in items. I.e. the Nth cartesian power of items. I.e. the cartesian product of items with itself N times.

check_all

Helper which calls check_lattice_ord, check_partial_ord_properties, check_lattice_properties, and check_lattice_bot.

check_atomize_each

Check that the atomized lattice points re-merge to form the same original lattice point, for each item in items.

check_lattice_bot

Checks that the item which is bot is less than (or equal to) all other items.

check_lattice_ord

Check that the lattice’s PartialOrd implementation agrees with the NaiveLatticeOrd partial order derived from Merge.

check_lattice_properties

Check lattice associativity, commutativity, and idempotence.

check_lattice_top

Checks that the item which is top is greater than (or equal to) all other items.

check_partial_ord_properties

Checks PartialOrd, PartialEq, and Eq’s reflexivity, symmetry, transitivity, and duality.