Trait lattices::IsBot

pub trait IsBot {
    // Required method
    fn is_bot(&self) -> bool;
}

Trait to check if a lattice instance is bottom (⊥).

Required Methods

fn is_bot(&self) -> bool

Returns if self is lattice bottom (⊥).

Must be consistent with equality, any element equal to bottom is also considered to be bottom.