lattices

Trait IsTop

pub trait IsTop {
    // Required method
    fn is_top(&self) -> bool;
}

Trait to check if a lattice instance is top (⊤) and therefore cannot change any futher.

Required Methods

fn is_top(&self) -> bool

Returns if self is lattice top (⊤).

Must be consistent with equality, any element equal to top is also considered to be top.