Type Alias lattices::union_find::UnionFindBTreeMap

pub type UnionFindBTreeMap<K> = UnionFind<BTreeMap<K, Cell<K>>>;

std::collections::BTreeMap-backed UnionFind lattice.

Aliased Type

struct UnionFindBTreeMap<K>(/* private fields */);

Implementations

impl<Map> UnionFind<Map>

pub fn new(val: Map) -> Self

Create a new UnionFind from a Map.

pub fn new_from(val: impl Into<Map>) -> Self

Create a new UnionFind from an Into<Map>.

pub fn as_reveal_ref(&self) -> &Map

Reveal the inner value as a shared reference.

pub fn as_reveal_mut(&mut self) -> &mut Map

Reveal the inner value as an exclusive reference.

pub fn into_reveal(self) -> Map

Gets the inner by value, consuming self.

impl<Map, K> UnionFind<Map>where Map: MapMut<K, Cell<K>, Key = K, Item = Cell<K>>, K: Copy + Eq,

pub fn union(&mut self, a: K, b: K) -> Min<bool>

Union the sets containg a and b.

Returns true if the sets changed, false if a and b were already in the same set. Once this returns false it will always return false for the same a and b, therefore it returns a Min<bool> lattice.

impl<MapSelf, K> UnionFind<MapSelf>where MapSelf: Map<K, Cell<K>, Key = K, Item = Cell<K>>, K: Copy + Eq,

pub fn same(&self, a: K, b: K) -> Max<bool>

Returns if a and b are in the same set.

This method is monotonic: once this returns true it will always return true for the same a and b, therefore it returns a Max<bool> lattice.