Trait lattices::LatticeFrom

pub trait LatticeFrom<Other> {
    // Required method
    fn lattice_from(other: Other) -> Self;
}

Same as From but for lattices.

This should only be implemented between different representations of the same lattice type. This should recursively convert nested lattice types, but not non-lattice (“scalar”) types.

Required Methods

fn lattice_from(other: Other) -> Self

Convert from the Other lattice into Self.