Trait lattices::LatticeOrd

pub trait LatticeOrd<Rhs = Self>: PartialOrd<Rhs> { }

Trait for lattice partial order comparison PartialOrd is implemented for many things, this trait can be used to require the type be a lattice.